move social link sharing component into common component
